Replace the old customer_area page with the new one, store the old one as customer_area_archive
