+++ packages/client/src/pages/customer_area/__tests__/CustomerAreaPage.test.tsx
