+++ packages/client/src/pages/customer_area/index.tsx
@@ -1,50 +1,103 @@
-import React, { useEffect, useMemo } from "react";
-import { Route, Switch, useParams } from "react-router-dom";
-import { useSelector } from "react-redux";
+import React, { useEffect, useState } from "react";
+import { DateTime } from "luxon";
+import { useParams } from "react-router-dom";
+import { useDispatch, useSelector } from "react-redux";
 
+import { CalendarNav, Layout, TabItem } from "@eisbuk/ui";
+import { Calendar, AccountCircle } from "@eisbuk/svg";
 import {
-  OrgSubCollection,
   BookingSubCollection,
   Collection,
+  OrgSubCollection,
 } from "@eisbuk/shared";
 
-import { CustomerRoute, Routes } from "@/enums/routes";
-
-import CustomerSlots from "./CustomerSlots";
-import BookingsCalendar from "./BookingsCalendar";
-import CustomerNavigation from "./CustomerNavigation";
-import AppbarCustomer from "@/components/layout/AppbarCustomer";
-import AppbarAdmin from "@/components/layout/AppbarAdmin";
+import BookView from "./views/Book";
+import CalendarView from "./views/Calendar";
+import { NotificationsContainer } from "@/features/notifications/components";
 
 import useFirestoreSubscribe from "@/react-redux-firebase/hooks/useFirestoreSubscribe";
-
-import {
-  getBookingsCustomer,
-  getBookedSlots,
-  getSlotsForCustomer,
-} from "@/store/selectors/bookings";
-import { getIsAdmin } from "@/store/selectors/auth";
+import { getBookingsCustomer } from "@/store/selectors/bookings";
 import { getCalendarDay } from "@/store/selectors/app";
+import { changeCalendarDate } from "@/store/actions/appActions";
 
-import { splitSlotsByCustomerRoute } from "./utils";
 import { setSecretKey, unsetSecretKey } from "@/utils/localStorage";
 
 /**
- * Customer sub routes:
- * - renders the apropriate `customerRoute` within `CustomersPage` and `CustomerNavigation`
- * - catches all `/customers/:secretKey/<customerRoute>` routes
- * - initializes `useFirestoreConnect` with appropriate params
- * - gets all slots and bookings from store (for appropriate timeframe)
- *   and processes to prepare `slots`/`subscribedSlots`/`bookings` props for each respective sub route
- * - renders appropriate sub route with respect to `customerRoute` provided
+ * Customer area page component
  */
 const CustomerArea: React.FC = () => {
-  const { customerRoute, secretKey } = useParams<{
+  useSecretKey();
+
+  // Subscribe to necessary collections
+  useFirestoreSubscribe([
+    OrgSubCollection.SlotsByDay,
+    OrgSubCollection.Bookings,
+    Collection.PublicOrgInfo,
+    BookingSubCollection.BookedSlots,
+    BookingSubCollection.Calendar,
+  ]);
+
+  const calendarNavProps = useDate();
+
+  // Get customer data necessary for rendering/functoinality
+  const customerData = useSelector(getBookingsCustomer);
+
+  // Get appropriate view to render
+  const views = {
+    BookView,
+    CalendarView,
+  };
+  const [view, setView] = useState<keyof typeof views>("BookView");
+  const additionalButtons = (
+    <>
+      <TabItem
+        Icon={Calendar as any}
+        label="Book"
+        onClick={() => setView("BookView")}
+        active={view === "BookView"}
+      />
+      <TabItem
+        Icon={AccountCircle as any}
+        label="Calendar"
+        onClick={() => setView("CalendarView")}
+        active={view === "CalendarView"}
+      />
+    </>
+  );
+  const CustomerView = views[view];
+
+  return (
+    <Layout
+      Notifications={NotificationsContainer}
+      additionalButtons={additionalButtons}
+      user={customerData}
+    >
+      <CalendarNav {...calendarNavProps} jump="month" />
+      <div className="content-container">
+        <div className="px-[44px] py-4">
+          <CustomerView />
+        </div>
+      </div>
+    </Layout>
+  );
+};
+
+/**
+ * Secret key logic abstracted away in a hook for easier readability
+ */
+const useSecretKey = () => {
+  // Secret key is provided as a route param to the customer_area page
+  const { secretKey } = useParams<{
     secretKey: string;
-    customerRoute: CustomerRoute;
   }>();
 
-  // store secret key to local storage
+  /**
+   * @TODO this disables the user (admin)
+   * from looking at bookings for multiple
+   * customers in different tabs and we should find
+   * a way around it (probably store the key in store)
+   */
+  // Store secretKey to localStorage
   // for easier access
   useEffect(() => {
     setSecretKey(secretKey);
@@ -55,88 +108,24 @@ const CustomerArea: React.FC = () => {
     };
   }, [secretKey]);
 
-  useFirestoreSubscribe([
-    OrgSubCollection.Bookings,
-    Collection.PublicOrgInfo,
-    BookingSubCollection.BookedSlots,
-    BookingSubCollection.Calendar,
-    OrgSubCollection.SlotsByDay,
-  ]);
+  return secretKey;
+};
 
-  const customerData = useSelector(getBookingsCustomer);
+/**
+ * Date logic abstracted away in a hook for readability.
+ * Reads current date from Redux store and handles updates of the current date.
+ * The returned structure can directly be passed as props to CalendarNav component
+ */
+const useDate = (): Pick<
+  Parameters<typeof CalendarNav>[0],
+  "date" | "onChange"
+> => {
+  const dispatch = useDispatch();
+
   const date = useSelector(getCalendarDay);
+  const onChange = (date: DateTime) => dispatch(changeCalendarDate(date));
 
-  // only the "calendar" will use "week" timeframe, the rest will use "month"
-  const timeframe = customerRoute === CustomerRoute.Calendar ? "week" : "month";
-
-  // get slots from store for apropriate timespan
-  const slotsSelector = useMemo(
-    () =>
-      customerData && date
-        ? getSlotsForCustomer(customerData.category, timeframe, date)
-        : () => ({}),
-    [date, customerData, timeframe]
-  );
-
-  // get raw slots (slots by day) from store
-  const rawSlots = useSelector(slotsSelector);
-
-  // process slots for each route
-  const {
-    [CustomerRoute.BookIce]: bookIceSlots,
-    [CustomerRoute.BookOffIce]: bookOffIceSlots,
-    [CustomerRoute.Calendar]: calendarSlots,
-  } = splitSlotsByCustomerRoute(rawSlots);
-
-  // create bookings to display
-  const bookedSlots = useSelector(getBookedSlots);
-
-  const isAdmin = useSelector(getIsAdmin);
-
-  const title = customerData
-    ? `${customerData.name} ${customerData.surname}`
-    : "";
-
-  const headers = (
-    <>
-      {isAdmin && <AppbarAdmin />}
-      <AppbarCustomer headingText={title} />
-    </>
-  );
-
-  return (
-    <>
-      {headers}
-      <CustomerNavigation />
-      <Switch>
-        <Route
-          path={`${Routes.CustomerArea}/:secretKey/${CustomerRoute.BookIce}`}
-        >
-          <CustomerSlots
-            view={CustomerRoute.BookIce}
-            slots={bookIceSlots}
-            rawSlots={rawSlots}
-            {...{ bookedSlots }}
-          />
-        </Route>
-        <Route
-          path={`${Routes.CustomerArea}/:secretKey/${CustomerRoute.BookOffIce}`}
-        >
-          <CustomerSlots
-            view={CustomerRoute.BookOffIce}
-            slots={bookOffIceSlots}
-            rawSlots={rawSlots}
-            {...{ bookedSlots }}
-          />
-        </Route>
-        <Route
-          path={`${Routes.CustomerArea}/:secretKey/${CustomerRoute.Calendar}`}
-        >
-          <BookingsCalendar slots={calendarSlots} {...{ bookedSlots }} />
-        </Route>
-      </Switch>
-    </>
-  );
+  return { date, onChange };
 };
 
 export default CustomerArea;
