+++ packages/client/src/pages/customer_area_archive/BookingsCalendar/__tests__/BookingsCalendar.test.tsx
