+++ packages/client/src/pages/customer_area_archive/CustomerSlots/CustomerSlots.stories.tsx
