+++ packages/client/src/pages/customer_area_archive/CustomerSlots/__tests__/CustomerSlots.test.tsx
